Renamed main modules
