+++ awesome/neconfig/user/look/widgets/user_look_titlebar_client_colors.lua
@@ -2,22 +2,23 @@
 --  █  █  █  █▄▄ ██▄ █▄█ █▀█ █▀▄   █▄▄ █▄▄ █ ██▄ █ ▀█  █    █▄▄ █▄█ █▄▄ █▄█ █▀▄ ▄█
 local titlebar_client_colors = {
     ['Libfm-pref-apps'] = '#121212',
-    ['qt5ct'] = '#222020',
-    ['Pcmanfm'] = '#272727',
-    ['firefox'] = '#1c1b22ff',
-    ['gammy'] = '#1b1c27',
-    ['Electron'] = '#2f3241',
-    ['Gimp-2.10'] = '#484848',
-    ['DarkPlaces'] = '#000000',
-    ['Inkscape'] = '#272727',
-    ['Nitrogen'] = '#121212',
-    ['Lxappearance'] = '#121212',
-    ['code-oss'] = '#272727',
     ['Qalculate-gtk'] = '#272727',
+    ['qt5ct'] = '#222020',
     ['com.github.joseexposito.touche'] = '#ebebebff',
+    ['Pcmanfm'] = '#272727',
+    ['gammy'] = '#1b1c27',
+    ['Gimp-2.10'] = '#484848',
     ['Alacritty'] = '#080808cc',
     ['discord'] = '#36393f',
-    ['feh'] = '#646464'
+    ['Lxappearance'] = '#121212',
+    ['Inkscape'] = '#272727',
+    ['code-oss'] = '#272727',
+    ['DarkPlaces'] = '#000000',
+    ['VSCodium'] = '#1e1e1e',
+    ['Nitrogen'] = '#121212',
+    ['firefox'] = '#1c1b22ff',
+    ['feh'] = '#646464',
+    ['Electron'] = '#2f3241'
 }
 
 return titlebar_client_colors
